# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Rating:
    properties:
      created_at:
        type: string
      episode_id:
        type: string
      id:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  usecase.CreateRatingInput:
    properties:
      rating:
        type: integer
    type: object
...
  usecase.EpisodeRatingStatsResult:
    properties:
      average_rating:
        type: number
      distribution:
        additionalProperties:
          type: integer
        type: object
      total_ratings:
        type: integer
    type: object
...
  usecase.MyRatingItem:
    properties:
      created_at:
        type: string
      episode:
        $ref: '#/definitions/usecase.RatingEpisodeInfo'
      id:
        type: string
      podcast:
        $ref: '#/definitions/usecase.RatingPodcastInfo'
      rating:
        type: integer
      updated_at:
        type: string
    type: object
  usecase.MyRatingListResult:
    properties:
      ratings:
        items:
          $ref: '#/definitions/usecase.MyRatingItem'
        type: array
      total:
        type: integer
    type: object
...
  usecase.PodcastRatingResult:
    properties:
      average_rating:
        type: number
      total_ratings:
        type: integer
    type: object
...
  usecase.RatingEpisodeInfo:
    properties:
      artwork_url:
        type: string
      id:
        type: string
      podcast_id:
        type: string
      title:
        type: string
    type: object
  usecase.RatingPodcastInfo:
    properties:
      artwork_url:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
  usecase.UpdateRatingInput:
    properties:
      rating:
        type: integer
    type: object
  usecase.UserRatingStatsResult:
    properties:
      average_rating:
        type: number
      distribution:
        additionalProperties:
          type: integer
        type: object
      total_ratings:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: お笑いラジオの視聴記録SNS API
  title: Podlog API
  version: "1.0"
paths:
...
  /episodes/{id}/ratings:
    get:
      description: エピソードの平均評価・総件数・星別分布を取得します
      parameters:
      - description: エピソードID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.EpisodeRatingStatsResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: エピソードの評価集計
      tags:
      - ratings
    post:
      consumes:
      - application/json
      description: 認証ユーザーがエピソードに評価（1〜5）を投稿します。既に投稿済みの場合は 409 を返します（FE は PUT /episodes/{id}/ratings/mine
        にフォールバック）。
      parameters:
      - description: エピソードID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 評価値
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateRatingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Rating'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 評価投稿
      tags:
      - ratings
  /episodes/{id}/ratings/mine:
    delete:
      description: 認証ユーザーの自分の評価のみを削除します（感想は別管理）。
      parameters:
      - description: エピソードID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 評価削除
      tags:
      - ratings
    get:
      description: 認証ユーザーが指定エピソードに投稿した評価を取得します。未投稿の場合は 404 を返します。
      parameters:
      - description: エピソードID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rating'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 自分の評価取得
      tags:
      - ratings
    put:
      consumes:
      - application/json
      description: 認証ユーザーの自分の評価を更新します
      parameters:
      - description: エピソードID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 評価値
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateRatingInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rating'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 評価更新
      tags:
      - ratings
...
  /podcasts/{id}/rating:
    get:
      description: ポッドキャストに紐づく全エピソードの評価から平均評価と総件数を集計します
      parameters:
      - description: ポッドキャストID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.PodcastRatingResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ポッドキャスト平均評価
      tags:
      - ratings
...
  /users/{username}/ratings/stats:
    get:
      description: ユーザー名を指定して評価の統計値（平均・総件数・星別分布）を取得します
      parameters:
      - description: ユーザー名
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.UserRatingStatsResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザーの評価統計サマリー（公開）
      tags:
      - ratings
...
  /users/me/ratings:
    get:
      description: 認証ユーザーが投稿した評価をエピソード・番組情報付きで一覧取得します（設定ページ等での確認用途）。
      parameters:
      - default: 20
        description: 最大取得件数
        in: query
        name: limit
        type: integer
      - default: 0
        description: スキップ件数
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.MyRatingListResult'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 自分の評価一覧
      tags:
      - ratings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"